requirements.txt
arrow==1.3.0
asttokens==2.4.1
awscli==1.34.25
binaryornot==0.4.4
black==24.8.0
bleach==6.1.0
botocore==1.35.25
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cookiecutter==2.6.0
cookiecutter-data-science==2.0.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
docutils==0.16
executing==2.1.0
feature-engine==1.8.1
flake8==7.1.1
fonttools==4.54.1
geodatasets==2024.8.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
ghp-import==2.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
kiwisolver==1.4.7
lightgbm==4.5.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.1.1
opendatasets==0.1.22
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
pyzmq==26.2.0
requests==2.32.3
rich==13.8.1
rsa==4.7.2
s3transfer==0.10.2
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.3
text-unidecode==1.3
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
tzdata==2024.1
urllib3==2.2.3
watchdog==5.0.2
wcwidth==0.2.13
webencodings==0.5.1
